Moves adHocConnect to the respective socketViews, Adds allowAdHoc prop to the socketAddAtom component Adds allowAdHoc prop to the atom-content-chats.jsx and atom-content-socket.jsx component
